pyproject.toml
[project]
name = "hijribot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "hijridate>=2.5.0",
    "pytest>=9.0.2",
    "python-telegram-bot>=22.5",
]
